selenium/python/requirements.txt
pytest-selenium
selenium>=4.32.0